chore : alert 폴더 삭제 및 컴포넌트, 변수명 수정
