+++ components/form/signupForm/index.tsx
@@ -9,7 +9,7 @@ import ImageUpload from "@/components/form/input/ImageUploadInput";
 import IntroTextarea from "@/components/form/input/IntroTextarea";
 import RadioInput from "@/components/form/input/RadioInput";
 import TagInput from "@/components/form/input/TagInput";
-import AlertModal from "@/components/modal/alert";
+import AlertModal from "@/components/modal";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { regEmail, regPassword } from "@/lib/utils/regexp";
@@ -46,14 +46,14 @@ function SignUpForm() {
   return (
     <>
       <form onSubmit={handleSubmit(onSubmit)}>
-        <div className="bg-white w-[956px] px-32 py-48 flex flex-col rounded-32 items-center tablet:w-[720px] mobile:w-[312px]">
+        <div className="flex w-[956px] flex-col items-center rounded-32 bg-white px-32 py-48 tablet:w-[720px] mobile:w-[312px]">
           <div className="flex flex-col gap-24">
-            <div className="flex items-center mb-20">
-              <div className="w-294 h-px bg-line-02 tablet:w-248 mobile:w-52"></div>
+            <div className="mb-20 flex items-center">
+              <div className="h-px w-294 bg-line-02 tablet:w-248 mobile:w-52"></div>
               <div className="text-18 tablet:text-16 mobile:text-14">
                 필수 정보 입력
               </div>
-              <div className="w-294 h-px bg-line-02 tablet:w-248 mobile:w-52"></div>
+              <div className="h-px w-294 bg-line-02 tablet:w-248 mobile:w-52"></div>
             </div>
             <div className="flex flex-col gap-8">
               <Label htmlFor="email">
@@ -62,17 +62,17 @@ function SignUpForm() {
               <Input
                 id="email"
                 type="email"
-                className={`w-[756px] h-52 bg-bg-02 placeholder:text-text-05 tablet:w-[672px] mobile:w-272 border border-line-02 rounded-12 px-16 focus-visible:ring-0 focus-visible:ring-offset-0 focus:bg-white focus:border focus:border-line-01 ${errors.email && "bg-input-error border-0"}`}
+                className={`h-52 w-[756px] rounded-12 border border-line-02 bg-bg-02 px-16 placeholder:text-text-05 focus:border focus:border-line-01 focus:bg-white focus-visible:ring-0 focus-visible:ring-offset-0 tablet:w-[672px] mobile:w-272 ${errors.email && "border-0 bg-input-error"}`}
                 placeholder="이메일을 입력해 주세요"
                 {...register("email", { required: true, pattern: regEmail })}
               />
               {errors.email && errors.email.type === "required" && (
-                <span className="text-system-error text-12">
+                <span className="text-12 text-system-error">
                   이메일을 입력해 주세요
                 </span>
               )}
               {errors.email && errors.email.type === "pattern" && (
-                <span className="text-system-error text-12">
+                <span className="text-12 text-system-error">
                   이메일 형식으로 작성해 주세요
                 </span>
               )}
@@ -84,7 +84,7 @@ function SignUpForm() {
               <Input
                 id="nickname"
                 type="text"
-                className={`w-[756px] h-52 bg-bg-02 placeholder:text-text-05 tablet:w-[672px] mobile:w-272 border border-line-02 rounded-12 px-16 focus-visible:ring-0 focus-visible:ring-offset-0 focus:bg-white focus:border focus:border-line-01 ${errors.nickname && "bg-input-error border-0"}`}
+                className={`h-52 w-[756px] rounded-12 border border-line-02 bg-bg-02 px-16 placeholder:text-text-05 focus:border focus:border-line-01 focus:bg-white focus-visible:ring-0 focus-visible:ring-offset-0 tablet:w-[672px] mobile:w-272 ${errors.nickname && "border-0 bg-input-error"}`}
                 placeholder="닉네임을 입력해 주세요"
                 {...register("nickname", {
                   required: true,
@@ -93,17 +93,17 @@ function SignUpForm() {
                 })}
               />
               {errors.nickname && errors.nickname.type === "required" && (
-                <span className="text-system-error text-12">
+                <span className="text-12 text-system-error">
                   닉네임을 입력해 주세요
                 </span>
               )}
               {errors.nickname && errors.nickname.type === "minLength" && (
-                <span className="text-system-error text-12">
+                <span className="text-12 text-system-error">
                   닉네임은 최소 2글자 입니다
                 </span>
               )}
               {errors.nickname && errors.nickname.type === "maxLength" && (
-                <span className="text-system-error text-12">
+                <span className="text-12 text-system-error">
                   닉네임은 최대 8글자 입니다
                 </span>
               )}
@@ -116,18 +116,18 @@ function SignUpForm() {
                 <Input
                   id="password"
                   type={passwordShown ? "text" : "password"}
-                  className={`w-[756px] h-52 bg-bg-02 placeholder:text-text-05 tablet:w-[672px] mobile:w-272 border border-line-02 rounded-12 px-16 focus-visible:ring-0 focus-visible:ring-offset-0 focus:bg-white focus:border focus:border-line-01 ${errors.password && "bg-input-error text-text-02 border-0"}`}
+                  className={`h-52 w-[756px] rounded-12 border border-line-02 bg-bg-02 px-16 placeholder:text-text-05 focus:border focus:border-line-01 focus:bg-white focus-visible:ring-0 focus-visible:ring-offset-0 tablet:w-[672px] mobile:w-272 ${errors.password && "border-0 bg-input-error text-text-02"}`}
                   placeholder="비밀번호를 입력해 주세요"
                   {...register("password", {
                     required: true,
                     pattern: regPassword,
                   })}
                 />
-                <div className="absolute inset-y-0 right-0 flex items-center justify-center px-16 h-full">
+                <div className="absolute inset-y-0 right-0 flex h-full items-center justify-center px-16">
                   <button
                     type="button"
                     onClick={togglePasswordVisiblity}
-                    className="w-24 h-24 relative"
+                    className="relative h-24 w-24"
                   >
                     <Image
                       src={`/icons/eye-${passwordShown ? "off" : "on"}.png`}
@@ -138,12 +138,12 @@ function SignUpForm() {
                 </div>
               </div>
               {errors.password && errors.password.type === "required" && (
-                <span className="text-system-error text-12">
+                <span className="text-12 text-system-error">
                   비밀번호를 입력해 주세요
                 </span>
               )}
               {errors.password && errors.password.type === "pattern" && (
-                <span className="text-system-error text-12">
+                <span className="text-12 text-system-error">
                   영어, 숫자, 특수문자를 조합하여 8자리 이상 입력해 주세요
                 </span>
               )}
@@ -156,18 +156,18 @@ function SignUpForm() {
                 <Input
                   id="confimPassword"
                   type={confirmPasswordShown ? "text" : "password"}
-                  className={`w-[756px] h-52 bg-bg-02 placeholder:text-text-05 tablet:w-[672px] mobile:w-272 border border-line-02 rounded-12 px-16 focus-visible:ring-0 focus-visible:ring-offset-0 focus:bg-white focus:border focus:border-line-01 ${errors.confirmPassword && "bg-input-error border-0"}`}
+                  className={`h-52 w-[756px] rounded-12 border border-line-02 bg-bg-02 px-16 placeholder:text-text-05 focus:border focus:border-line-01 focus:bg-white focus-visible:ring-0 focus-visible:ring-offset-0 tablet:w-[672px] mobile:w-272 ${errors.confirmPassword && "border-0 bg-input-error"}`}
                   placeholder="비밀번호를 다시 입력해 주세요"
                   {...register("confirmPassword", {
                     required: true,
                     validate: value => value === password,
                   })}
                 />
-                <div className="absolute inset-y-0 right-0 flex items-center justify-center px-16 h-full">
+                <div className="absolute inset-y-0 right-0 flex h-full items-center justify-center px-16">
                   <button
                     type="button"
                     onClick={toggleConfirmPasswordVisibility}
-                    className="w-24 h-24 relative"
+                    className="relative h-24 w-24"
                   >
                     <Image
                       src={`/icons/eye-${confirmPasswordShown ? "off" : "on"}.png`}
@@ -178,7 +178,7 @@ function SignUpForm() {
                 </div>
               </div>
               {errors.confirmPassword && (
-                <span className="text-system-error text-12">
+                <span className="text-12 text-system-error">
                   비밀번호가 일치하지 않습니다
                 </span>
               )}
@@ -217,14 +217,14 @@ function SignUpForm() {
               />
             </div>
           </div>
-          <div className="flex items-center my-40 gap-32 tablet:gap-28">
-            <div className="w-294 h-px bg-line-02 tablet:w-248 mobile:w-52"></div>
+          <div className="my-40 flex items-center gap-32 tablet:gap-28">
+            <div className="h-px w-294 bg-line-02 tablet:w-248 mobile:w-52"></div>
             <div className="text-18 tablet:text-16 mobile:text-14">
               추가 정보 입력
             </div>
-            <div className="w-294 h-px bg-line-02 tablet:w-248 mobile:w-52"></div>
+            <div className="h-px w-294 bg-line-02 tablet:w-248 mobile:w-52"></div>
           </div>
-          <div className="flex flex-col gap-24 items-center">
+          <div className="flex flex-col items-center gap-24">
             <Controller
               control={control}
               name="imageUpload"
@@ -264,11 +264,11 @@ function SignUpForm() {
             </div>
           </div>
         </div>
-        <div className="w-[956px] mt-40 flex flex-col rounded-32 items-center tablet:w-[720px] mobile:w-[312px]">
+        <div className="mt-40 flex w-[956px] flex-col items-center rounded-32 tablet:w-[720px] mobile:w-[312px]">
           <button
             type="submit"
             disabled={!isValid}
-            className="w-240 h-52 bg-primary hover:bg-primary-press text-white rounded-32 flex justify-center items-center disabled:bg-line-02 disabled:text-text-04 mobile:w-180 mobile:h-44"
+            className="flex h-52 w-240 items-center justify-center rounded-32 bg-primary text-white hover:bg-primary-press disabled:bg-line-02 disabled:text-text-04 mobile:h-44 mobile:w-180"
           >
             가입하기
           </button>
@@ -276,7 +276,7 @@ function SignUpForm() {
       </form>
       {isModalOpen && (
         <AlertModal
-          alertType="signupSuccess"
+          modalType="signupSuccess"
           onClose={() => {
             setIsModalOpen(false);
           }}
