+++ components/modal/button.tsx
@@ -1,34 +1,31 @@
-import { AlertType } from "@/components/modal/alert";
+import { ModalType } from "@/components/modal";
 import { Button } from "@/components/ui/button";
 
-interface AlertModalButtonProps {
-  alertType: AlertType;
+interface ModalButtonProps {
+  modalType: ModalType;
   onClose: () => void;
 }
 
-export default function AlertModalButton({
-  alertType,
-  onClose,
-}: AlertModalButtonProps) {
+export default function ModalButton({ modalType, onClose }: ModalButtonProps) {
   let filledStyle: string = "";
   let ghostStyle: string = "";
   let filledText: string = "";
   let ghostText: string = "";
 
   if (
-    alertType === "emailNotFound" ||
-    alertType === "passwordMismatch" ||
-    alertType === "signupSuccess" ||
-    alertType === "writingSuccess" ||
-    alertType === "emailInUse" ||
-    alertType === "userProfile"
+    modalType === "emailNotFound" ||
+    modalType === "passwordMismatch" ||
+    modalType === "signupSuccess" ||
+    modalType === "writingSuccess" ||
+    modalType === "emailInUse" ||
+    modalType === "userProfile"
   ) {
-    filledStyle = `text-18 h-52 mobile:h-44  ${alertType === "userProfile" ? "w-240 mobile:w-full" : "w-full"}`;
+    filledStyle = `text-18 h-52 mobile:h-44  ${modalType === "userProfile" ? "w-240 mobile:w-full" : "w-full"}`;
     filledText = "확인";
   } else if (
-    alertType === "writingCancel" ||
-    alertType === "travelApply" ||
-    alertType === "travelCancle"
+    modalType === "writingCancel" ||
+    modalType === "travelApply" ||
+    modalType === "travelCancle"
   ) {
     filledStyle = "w-full text-18 h-52 mobile:h-44";
     ghostStyle = "w-full text-18 h-52 mobile:h-44";
